### 2017-12-15 #510 @whedon
[PRE REVIEW]: MixEst: An Estimation Toolbox for Mixture Models
**Submitting author:** @reshadh (<a href="http://orcid.org/0000-0002-3669-760X">Reshad Hosseini</a>)
**Repository:** <a href="https://github.com/utvisionlab/mixest" target ="_blank">https://github.com/utvisionlab/mixest</a>
**Version:** v1.2.0
