### 2017-12-15 #510 @arfon
Looks like @whedon doesn't quite know how to compile this paper. I've manually done it here [10.21105.joss.00510.pdf](https://github.com/openjournals/joss-reviews/files/1562585/10.21105.joss.00510.pdf)
