### 2017-12-20 #510 @arokem
👋 @josephhardinee: are you available to review this submission for JOSS?
