### 2018-01-02 #510 @josephhardinee
I am. Sorry I was out for holidays. This will be the first one I've reviewed so I'll probably require a little bit of hand holding.
