### 2018-01-07 #510 @arokem
@josephhardinee : thanks for getting back to me on this. No problem with this being your first review. I can help along!

I will start by assigning you as an editor and starting the review process. Once that's done, it will create a new issue for the review, with a check list of items that we'd like you to take a look at. The best way to help authors make it through the review is to post issues to the Github of their project, relating them to the review check list.  And the review issue gives you the opportunity to comment on other general aspects of the software that are not covered by the checklist -- you can definitely post issues about these as well.
