### 2018-02-10 #524 @arokem
@josephhardinee : just checking in on this. Have you had a chance to take a look at the software at all? Please let me know if you need any more guidance on how to go about this review.
