### 2018-02-13 #524 @josephhardinee
@arokem Sorry, I got part way through then got hit with a bunch of work deadlines. I will finish this review tonight.
