### 2018-02-23 493362f
some fixes
