+++ paper/paper.bib
@@ -1,145 +1,0 @@
-%% This BibTeX bibliography file was created using BibDesk.
-%% http://bibdesk.sourceforge.net/
-
-
-%% Created for Reshad Hosseini at 2015-07-17 10:02:19 +0430 
-
-
-%% Saved with string encoding Unicode (UTF-8) 
-
-
-
-@article{fraley1999mclust,
-	Author = {Fraley, Chris and Raftery, Adrian E},
-	Date-Added = {2015-07-17 05:32:15 +0000},
-	Date-Modified = {2015-07-17 05:32:15 +0000},
-	Journal = {Journal of Classification},
-	Number = {2},
-	Pages = {297--306},
-	Publisher = {Springer},
-	Title = {MCLUST: Software for model-based cluster analysis},
-	Volume = {16},
-	Year = {1999}}
-
-@article{biernacki2006model,
-	Author = {Biernacki, Christophe and Celeux, Gilles and Govaert, G{\'e}rard and Langrognet, Florent},
-	Date-Added = {2015-07-17 05:25:12 +0000},
-	Date-Modified = {2015-07-17 05:27:23 +0000},
-	Journal = {Computational Statistics and Data Analysis},
-	Number = {2},
-	Pages = {587--600},
-	Publisher = {Elsevier},
-	Title = {Model-based cluster and discriminant analysis with the MIXMOD software},
-	Volume = {51},
-	Year = {2006}}
-
-@article{Leisch_2004b,
-	Abstract = {{FlexMix implements a general framework for fitting discrete mixtures of regression models in the R statistical computing environment: three variants of the EM algorithm can be used for parameter estimation, regressors and responses may be multivariate with arbitrary dimension, data may be grouped, e.g., to account for multiple observations per individual, the usual formula interface of the S language is used for convenient model specification, and a modular concept of driver functions allows to interface many different types of regression models. Existing drivers implement mixtures of standard linear models, generalized linear models and model-based clustering. FlexMix provides the E-step and all data handling, while the M-step can be supplied by the user to easily define new models.}},
-	Author = {Leisch, Friedrich},
-	Date-Added = {2015-07-17 05:22:47 +0000},
-	Date-Modified = {2015-07-17 05:23:04 +0000},
-	Journal = {Journal of Statistical Software},
-	Number = {8},
-	Pages = {1--18},
-	Title = {{FlexMix}: A general framework for finite mixture models and latent class regression in {R}},
-	Volume = {11},
-	Year = {2004},
-	Bdsk-Url-1 = {http://www.jstatsoft.org/v11/i08/}}
-
-@article{Murphy01thebayes,
-	Author = {Kevin P. Murphy},
-	Date-Added = {2015-07-17 05:18:22 +0000},
-	Date-Modified = {2015-07-17 05:19:15 +0000},
-	Journal = {Computing Science and Statistics},
-	Pages = {2001},
-	Title = {The {Bayes Net} Toolbox for MATLAB},
-	Volume = {33},
-	Year = {2001}}
-
-@article{scikit-learn,
-	Author = {Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V. and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P. and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
-	Date-Added = {2015-07-17 05:14:37 +0000},
-	Date-Modified = {2015-07-17 05:14:49 +0000},
-	Journal = {Journal of Machine Learning Research},
-	Pages = {2825--2830},
-	Title = {Scikit-learn: Machine Learning in Python},
[…]
