+++ paper/paper.md
@@ -1,223 +1,0 @@
----
-title: 'MixEst: An Estimation Toolbox for Mixture Models'
-tags:
-  - mixture models 
-  - mixtures of experts 
-  - manifold optimization
-  - expectation-maximization 
-  - stochastic optimization
-authors:
- - name: Reshad Hosseini
-   orcid: 0000-0002-3669-760X
-   affiliation: 1
- - name: Mohamadreza Mash'al
-   affiliation: 1
-affiliations:
- - name: School of ECE, College of Engineering, University of Tehran
-   index: 1
-date: 15 December 2017
-bibliography: paper.bib
----
-
-# Summary #
-    
-Mixture models are powerful statistical models used in many applications ranging from density estimation to clustering and classification.
-When dealing with mixture models, there are many issues that the experimenter should be aware of and needs to solve.
-The MixEst toolbox is a powerful and user-friendly package for MATLAB that implements several state-of-the-art approaches to address these problems.
-Additionally, MixEst gives the possibility of using manifold optimization for fitting the density model, a feature specific to this toolbox.
-MixEst simplifies using and integration of mixture models in statistical models and applications.
-For developing mixture models of new densities, the user just needs to provide a few functions for that statistical distribution and the toolbox takes care of all the issues regarding mixture models.
-
-# Introduction #
-
-Mixture models are an integrated and fundamental component in many
-machine learning problems ranging from clustering to regression and
-classification [@McLPee00]. Estimating the parameters of mixture models
-is a challenging task due to the need to solve the following issues in
-mixture modeling:
-
--   Unboundedness of the likelihood: This problem occurs when one
-    component gets a small number of data points and its likelihood
-    becomes infinite [@ciuperca_penalized_2003].
-
--   Local maxima: The log-likelihood objective function for estimating
-    the parameters of mixture models is non-concave and has many local
-    maxima [@ueda_split_2000].
-
--   Correct number of components: In many applications, it is needed to
-    find the correct number of components [@khalili_variable_2007].
-
-Addressing these issues for a mixture density when it is not available
-in common mixture modeling toolboxes will cost a lot of time and effort
-for the experimenter. MixEst addresses all these issues not only for
-already implemented densities, but also for densities that the user may
-implement. By implementing densities, we mean implementing a few simple
-functions which will be briefly discussed in [Model Development Section](#model-development).
-
-This toolbox provides a framework for applying manifold optimization for
-estimating the parameters of mixture models. This is an important
-feature of this toolbox, because recent empirical evidence shows that
-manifold optimization can surpass expectation maximization in the case
-of mixtures of Gaussians [@hosseini2015]. It also opens the door for
-large-scale optimization by using stochastic optimization methods.
-Stochastic optimization also allows solving the likelihood unboundedness
-problem mentioned above, without the need of implementing a penalizing
-function for the parameters of the density.
-
-While several libraries are available for working with mixture models,
-to the best of our knowledge, none of them offers a modular and flexible
-framework that allows for fine-tuning the model structure or can provide
-universal algorithms for estimating model parameters solving all the
-problems listed above. A review of features available in some libraries
-can be seen in [Comparison
-Section](#feature-comparison).
-
-In the next section, we give a short overview of the toolbox and its
-features.
-
-# About the MixEst Toolbox # 
-
-This toolbox offers methods for constructing and estimating mixtures for
-joint density and conditional density modeling, therefore it is
-applicable to a wide variety of applications like clustering, regression
[…]
