+++ examples/example4.m
@@ -43,4 +43,4 @@ options.tolCostDiff = 1e-3;
 
 
 % run
-splitTmerge(data, 7, options)
+cem(data, options)
