+++ examples/example2.m
@@ -40,7 +40,7 @@ options.plotcost = true;
 options.verbosity = 2;
 options.solver = 'lbfgs';
 options.tolcostdiff = 1e-6;
-%options.crossval = true;
+options.crossval = true;
 options.minibatch.size = 10;
 options.maxiter = 200;
 options.minibatch.discardhistory = false;
