+++ examples/example1.m
@@ -30,16 +30,11 @@ options.plotGradNorm.axes = subplot(2,2,4);
 
 % main options
 options.verbosity = 2;
-options.solver = 'sgd';
-options.sgd.stepsize = 1;
-batchnum = size(data,2)/(2*5);
-it = 20;
-options.sgd.batchnum = batchnum;
-options.sgd.base = 10^(-3/batchnum/it);
+options.solver = 'lbfgs';
 % options.crossVal = true;
 % options.crossVal.tolIter = 100;
 % options.tolCostDiff = -Inf;
-options.penalize =  true;
+% options.penalize = true;
 % options.minibatch.iter = 5;
 % options.minibatch.discardHistory = false;
 % options.minibatch.size = 300;
