### 2018-02-23 518bbe3
some doc fixes from the past
