### 2018-03-25 #524 @arokem
Hi @josephhardinee : I saw that you posted an issue on the software repo. Do you have any more comments here? Several of the check-boxes are currently unchecked -- does that mean that the authors need to fix some of these things? Thanks!
