### 2018-04-08 #524 @arokem
Any news here? I see that many of the check marks are marked. Any other comments, so that authors can fill the other requirements?
