### 2018-04-22 #524 @arfon
👋@reshadh - I noticed that the paper has been withdrawn on the JOSS site. Did you do this in error or are you wanting to withdraw your paper?
