### 2018-04-23 #524 @arfon
> I actually withdrew the paper. Sorry, I thought the reviewers would be informed on this. Otherwise, I would have mentioned this here.

OK thanks for letting us know @reshadh.

@arokem @josephhardinee - I'm sorry that this paper isn't proceeding, especially given all of your work here @josephhardinee.
